requirements.txt
click 
pycryptodomex
mutagen
requests
deezer-py
spotipy>=2.11.0
py-cord
requests
bs4
pillow
